+++ packages/core/src/grid/Grid.tsx
@@ -24,15 +24,22 @@ export const Grid: FC<GridProps> = ({gutter = 20, columns = 12, layout = columns
 
   return (
     <Flex wrap="wrap">
-      {Children.map(elements, (child: ReactNode, index: number) => (
-        <FlexItem
-          key={index}
-          width={`calc((100% + ${gutter}px) * ${layoutArr[index % layoutLength] / columns} - ${gutter}px)`}
-          mt={index >= rowBlocksCount ? `${gutter}px` : 0}
-          ml={index % rowBlocksCount !== 0 ? `${gutter}px` : 0}
-          children={child}
-        />
-      ))}
+      {Children.map(elements, (child: ReactNode, index: number) => {
+        const oneColumnPercent = layoutArr[index % layoutLength] / columns
+        const width = gutter === 0
+          ? `${100 * oneColumnPercent}%`
+          : `calc(${100 * oneColumnPercent}% + ${gutter * oneColumnPercent - gutter}px)`
+
+        return (
+          <FlexItem
+            key={index}
+            width={width}
+            mt={index >= rowBlocksCount ? `${gutter}px` : 0}
+            ml={index % rowBlocksCount !== 0 ? `${gutter}px` : 0}
+            children={child}
+          />
+        )
+      })}
     </Flex>
   )
 }