+++ packages/core/src/menu/MenuControlState.ts
@@ -1,0 +1,6 @@
+import {RefObject} from 'react'
+
+export default interface SuggestControlState {
+  refs: RefObject<HTMLDivElement>[]
+  focused: number | undefined
+}