+++ packages/mobile/src/input-modal/index.ts
@@ -1,0 +1,5 @@
+import InputModal from './InputModal'
+
+export {
+  InputModal,
+}